+++ namespaces/Player.cpp
@@ -8,6 +8,6 @@ namespace Player
 void getEntity(CScriptVar* jsfunc, void*)
 {
 	// TODO: this doesn't work since CScriptVar doesn't have a uint64_t constructor
-	jsfunc->setReturnVar(new CScriptVar(Entity$getUniqueID(MCPE_localplayer)));
+	jsfunc->setReturnVar(new CScriptVar((int) Entity$getUniqueID(MCPE_localplayer)));
 }
 };
